Corrected gitignore. Compendium style format
Revised repository in a Compendium style format like TTRPG Cli  Tool for DnD5e.
Separation of fixed game rules and expandable "items" like classes, spells, gear, weapons, etc...
The idea is to have a repo with only the rules and items to be downloaded and copied to any obsidian vault.